xpcom: warning
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@48708 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/libs/xpcom18a4/python/src/VariantUtils.cpp
@@ -1116,7 +1116,7 @@ PRBool PyXPCOM_InterfaceVariantHelper::Init(PyObject *obParams)
 	if (PySequence_Length(m_pyparams) != total_params_needed) {
 #ifdef VBOX
                 PyErr_Format(PyExc_ValueError, "The type descriptions indicate %d args are needed, but %ld were provided",
-			total_params_needed, PySequence_Length(m_pyparams));
+			total_params_needed, (long)PySequence_Length(m_pyparams));
 #else
 		PyErr_Format(PyExc_ValueError, "The type descriptions indicate %d args are needed, but %d were provided",
 			total_params_needed, PySequence_Length(m_pyparams));
